refactor: Moving specvol_sso_0 to gsw_internal_funcs
