+++ src/lib.rs
@@ -29,6 +29,5 @@ pub mod volume;
 mod error;
 pub use crate::error::{Error, Result};
 
-pub use crate::volume::{
-    alpha, beta, rho, specvol, specvol_alpha_beta, specvol_anom_standard, specvol_sso_0,
-};
+pub use crate::gsw_internal_funcs::specvol_sso_0;
+pub use crate::volume::{alpha, beta, rho, specvol, specvol_alpha_beta, specvol_anom_standard};
