+++ src/volume.rs
@@ -53,7 +53,7 @@ fn non_dimensional_sa(sa: f64) -> Result<f64> {
 /// let p = 3812;
 /// assert!((3812.0 * 1e-4) > (3812.0 / 1e4))
 /// ```
-fn non_dimensional_p(p: f64) -> f64 {
+pub(crate) fn non_dimensional_p(p: f64) -> f64 {
     if cfg!(feature = "compat") {
         p * 1e-4
     } else {
@@ -173,53 +173,6 @@ pub fn specvol(sa: f64, ct: f64, p: f64) -> Result<f64> {
                     + z * (V004 + xs * V104 + ys * V014 + z * (V005 + z * V006))))))
 }
 
-/// Specific Volume of Standard Ocean Salinity and CT=0
-///
-/// This function calculates specific volume at the Standard Ocean Salinity,
-/// SSO, and at a Conservative Temperature of zero degrees C, as a function
-/// of pressure, p, in dbar, using a streamlined version of the 75-term CT
-/// version of specific volume, that is, a streamlined version of the code
-/// "specvol(SA,CT,p)".
-///
-/// version: 3.06.12
-///
-/// If using compat (truncated constants) there is a difference of O[1e-19],
-/// which is negligible but enough to fail the validation tests.
-pub fn specvol_sso_0(p: f64) -> f64 {
-    const VXX0: f64 = if cfg!(feature = "compat") {
-        9.726_613_854_843_87e-4
-    } else {
-        9.726_613_854_843_871e-4
-    };
-
-    const VXX1: f64 = if cfg!(feature = "compat") {
-        -4.505_913_211_160_929e-5
-    } else {
-        -4.505_913_211_160_931e-5
-    };
-
-    const VXX2: f64 = if cfg!(feature = "compat") {
-        7.130_728_965_927_127e-6
-    } else {
-        7.130_728_965_927_128e-6
-    };
-
-    const VXX3: f64 = if cfg!(feature = "compat") {
-        -6.657_179_479_768_312e-7
-    } else {
-        -6.657_179_479_768_313e-7
-    };
-    const VXX4: f64 = if cfg!(feature = "compat") {
-        -2.994_054_447_232_88e-8
-    } else {
-        -2.994_054_447_232_877_6e-8
-    };
-
-    let p = non_dimensional_p(p);
-
-    VXX0 + p * (VXX1 + p * (VXX2 + p * (VXX3 + p * (VXX4 + p * (V005 + V006 * p)))))
-}
-
 /// Specific Volume Anomaly of Standard Ocean Salinity and CT=0
 ///
 /// Specific volume anomaly with reference of SA = SSO & CT = 0 (75-term equation)
@@ -235,7 +188,7 @@ pub fn specvol_sso_0(p: f64) -> f64 {
 /// * `specvol_anom`: specific volume anomaly of seawater [m3 kg-1]
 ///
 pub fn specvol_anom_standard(sa: f64, ct: f64, p: f64) -> Result<f64> {
-    Ok(specvol(sa, ct, p)? - specvol_sso_0(p))
+    Ok(specvol(sa, ct, p)? - crate::gsw_internal_funcs::specvol_sso_0(p))
 }
 
 pub fn specvol_first_derivatives(sa: f64, ct: f64, p: f64) -> Result<(f64, f64, f64)> {
@@ -502,7 +455,7 @@ pub fn sigma4(sa: f64, ct: f64) -> Result<f64> {
 
 #[cfg(test)]
 mod tests {
-    use super::{alpha, beta, specvol, specvol_anom_standard, specvol_sso_0, GSW_SSO};
+    use super::{alpha, beta, specvol, specvol_anom_standard, GSW_SSO};
 
     #[test]
     // Test value from Roquet 2015, Appendix C.3
@@ -522,17 +475,6 @@ mod tests {
                     .abs()
                     < f64::EPSILON
             );
-        }
-    }
-
-    /// specvol() at SSO & CT=0 should be identical to specvol_sso_0()
-    #[test]
-    fn test_specvol_vs_specvol_sso_0() {
-        let p_to_test: [f64; 5] = [0., 10., 100., 1000., 5000.];
-        for p in p_to_test.iter().cloned() {
-            let specvol = specvol(GSW_SSO, 0., p).unwrap();
-            let specvol_sso_0 = specvol_sso_0(p);
-            assert!((specvol - specvol_sso_0).abs() < f64::EPSILON);
         }
     }
